librados, osd: Integrate librados-dmclock using QoS profile manager
The original changes are adapted from the commits of
PR: https://github.com/ceph/ceph/pull/20235 to work with the
current implementation of mClock scheduler.

Change description:
Introduce the QosProfileMgr that manages a set of QosProfiles. Each
QosProfile is associated with a client (IoCtx), and encapsulates the
ServiceTracker object. The QosProfileMgr provides interfaces to create,
set and release QoS profiles. The librados layer exposes C/C++ APIs to
create, set and release the client specific QoS profiles using an
opaque pointer to the QosProfileMgr. The QoS parameters like reservation,
weight, limit and the profile id are set within the QosProfile before
sending the op for the first time. The profile id is unique for a given
client and is incremented for any newly added client. The QosProfile is
re-used for a given client until it is released.

The Objecter is modified to use the QosProfile associated with the
ObjectOperation to inject the client specific QoS request parameters to
each MOSDOp before sending it. The Objecter also reads the QoS phase and
cost from the MOSDOpReply received from the OSD and feeds it back to the
QosProfile's ServiceTracker to calculate the new delta and rho parameters
for the next cycle of ops from the client. This removes the need to
maintain the service tracker object in the Objecter.

The original commits from where the changes are adapted, their authors and
description are mentioned below.

1. 38e249160ddc9dee73c145886857e15a42ea8403
   Author(s): J. Eric Ivancich <ivancich@redhat.com>
   Description: Work on dmclock-librados integration.

Adapted by:
Signed-off-by: Sridhar Seshasayee <sseshasa@redhat.com>

Original Author(s):
Signed-off-by: J. Eric Ivancich <ivancich@redhat.com>